Add signup client for error
